Manually adding subject array to object array
Noticed that adding "subject" as a string directly into the object array produced an unexpected strange result. The randomized strings in the subject array would ONLY appear in printed the object array IF it was also the same subject! However, that's kinda nice, so I'm leaving that in, while I have manually just added the subject strings to the object array.